move shared widget validation utils and constants to shared folder
