+++ app/client/src/pages/Editor/Explorer/JSDependencies.tsx
@@ -6,6 +6,7 @@ import { BindingText } from "pages/Editor/APIEditor/CommonEditorForm";
 import { extraLibraries } from "utils/DynamicBindingUtils";
 import CollapseToggle from "./Entity/CollapseToggle";
 import Collapse from "./Entity/Collapse";
+import { extraLibraries } from "@shared/ast";
 
 const Wrapper = styled.div`
   font-size: 14px;
