+++ app/client/src/utils/autocomplete/EntityDefinitions.ts
@@ -685,53 +685,6 @@ export const GLOBAL_DEFS = {
   },
 };
 
-export const GLOBAL_FUNCTIONS = {
-  "!name": "DATA_TREE.APPSMITH.FUNCTIONS",
-  navigateTo: {
-    "!doc": "Action to navigate the user to another page or url",
-    "!type":
-      "fn(pageNameOrUrl: string, params: {}, target?: string) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  showAlert: {
-    "!doc": "Show a temporary notification style message to the user",
-    "!type": "fn(message: string, style: string) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  showModal: {
-    "!doc": "Open a modal",
-    "!type": "fn(modalName: string) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  closeModal: {
-    "!doc": "Close a modal",
-    "!type": "fn(modalName: string) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  storeValue: {
-    "!doc": "Store key value data locally",
-    "!type": "fn(key: string, value: any) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  download: {
-    "!doc": "Download anything as a file",
-    "!type":
-      "fn(data: any, fileName: string, fileType?: string) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  copyToClipboard: {
-    "!doc": "Copy text to clipboard",
-    "!type": "fn(data: string, options: object) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  resetWidget: {
-    "!doc": "Reset widget values",
-    "!type":
-      "fn(widgetName: string, resetChildren: boolean) -> +Promise[:t=[!0.<i>.:t]]",
-  },
-  setInterval: {
-    "!doc": "Execute triggers at a given interval",
-    "!type": "fn(callback: fn, interval: number, id?: string) -> void",
-  },
-  clearInterval: {
-    "!doc": "Stop executing a setInterval with id",
-    "!type": "fn(id: string) -> void",
-  },
-};
-
 export const getPropsForJSActionEntity = ({
   config,
   data,
