+++ app/shared/ast/index.ts
@@ -15,14 +15,30 @@ import {
   MemberExpressionData,
 } from './src/index';
 
+import { ExtraLibrary } from './src/constants';
+
 // constants
-import { ECMA_VERSION, SourceType, NodeTypes } from './src/constants';
+import {
+  ECMA_VERSION,
+  SourceType,
+  NodeTypes,
+  extraLibraries,
+  extraLibrariesNames,
+  GLOBAL_FUNCTIONS,
+  GLOBAL_WORKER_SCOPE_IDENTIFIERS,
+  JAVASCRIPT_KEYWORDS,
+} from './src/constants';
 
 // JSObjects
 import { parseJSObjectWithAST } from './src/jsObject';
 
 // types or intefaces should be exported with type keyword, while enums can be exported like normal functions
-export type { ObjectExpression, PropertyNode, MemberExpressionData };
+export type {
+  ObjectExpression,
+  PropertyNode,
+  MemberExpressionData,
+  ExtraLibrary,
+};
 
 export {
   isIdentifierNode,
@@ -40,4 +56,9 @@ export {
   ECMA_VERSION,
   SourceType,
   NodeTypes,
+  extraLibraries,
+  GLOBAL_FUNCTIONS,
+  GLOBAL_WORKER_SCOPE_IDENTIFIERS,
+  extraLibrariesNames,
+  JAVASCRIPT_KEYWORDS,
 };
