+++ app/shared/ast/package.json
@@ -25,7 +25,10 @@
     "lodash": "^4.17.21",
     "rollup": "^2.77.0",
     "typescript": "4.5.5",
-    "unescape-js": "^1.1.4"
+    "unescape-js": "^1.1.4",
+    "moment-timezone": "^0.5.35",
+    "fast-xml-parser": "^3.17.5",
+    "node-forge": "^1.3.0"
   },
   "devDependencies": {
     "@babel/preset-typescript": "^7.17.12",
@@ -37,7 +40,9 @@
     "rollup-plugin-generate-package-json": "^3.2.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-typescript2": "^0.32.0",
-    "ts-jest": "27.0.0"
+    "ts-jest": "27.0.0",
+    "@types/moment-timezone": "^0.5.10",
+    "@types/node-forge": "^0.10.0"
   },
   "author": "",
   "license": "ISC"
