+++ app/shared/ast/src/utils.ts
@@ -1,11 +1,11 @@
+import {
+  extraLibrariesNames,
+  GLOBAL_FUNCTIONS,
+  GLOBAL_WORKER_SCOPE_IDENTIFIERS,
+  JAVASCRIPT_KEYWORDS,
+} from 'constants/global';
 import { has } from 'lodash';
 import unescapeJS from 'unescape-js';
-import {
-  WINDOW_OBJECT_METHODS,
-  WINDOW_OBJECT_PROPERTIES,
-  GLOBAL_SCOPE_OBJECTS,
-  JAVASCRIPT_KEYWORDS,
-} from './constants';
 
 const beginsWithLineBreakRegex = /^\s+|\s+$/;
 
@@ -17,11 +17,23 @@ export function sanitizeScript(js: string, evaluationVersion: number) {
   return evaluationVersion > 1 ? trimmedJS : unescapeJS(trimmedJS);
 }
 
-export function isInvalidEntiyName(name: string) {
+/**
+ *
+ * @param name
+ * @returns Checks if an identifier is a valid reference to an entity
+ * @example For binding {{function(){  
+ * const error = new Error("test error")
+ * return error.message
+}()}}
+Identifier "Error" is not a valid entity reference, as it is part of the identifiers globally available 
+in the worker context (where evaluations are done)
+
+ */
+export function isInvalidEntityReference(identifier: string) {
   return (
-    has(JAVASCRIPT_KEYWORDS, name) ||
-    has(GLOBAL_SCOPE_OBJECTS, name) ||
-    has(WINDOW_OBJECT_PROPERTIES, name) ||
-    has(WINDOW_OBJECT_METHODS, name)
+    has(JAVASCRIPT_KEYWORDS, identifier) ||
+    has(GLOBAL_WORKER_SCOPE_IDENTIFIERS, identifier) ||
+    extraLibrariesNames.includes(identifier) ||
+    has(GLOBAL_FUNCTIONS, identifier)
   );
 }
